requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
pydantic
python-dotenv
argon2-cffi
python-multipart
opencv-python
pytesseract
huggingface-hub
transformers
torch
requests
python-dotenv
requests
requests
python-dotenv
PyJWT